+++ Tiny.RestClient/Exceptions/SerializeException.cs
@@ -1,0 +1,16 @@
+﻿using System;
+
+namespace Tiny.RestClient
+{
+    /// <summary>
+    /// Class SerializeException.
+    /// </summary>
+    /// <seealso cref="TinyRestClientException" />
+    public class SerializeException : TinyRestClientException
+    {
+        internal SerializeException(Type typeToSerialiser, Exception innerException)
+            : base($"Unable to serialize {typeToSerialiser.Name}", innerException)
+        {
+        }
+    }
+}