+++ Tiny.RestClient/Listener/IListener.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Net.Http;
+using System.Threading.Tasks;
 
 namespace Tiny.RestClient
 {
@@ -19,7 +20,8 @@ namespace Tiny.RestClient
         /// <param name="uri">uri of the request</param>
         /// <param name="httpMethod">verb of the request</param>
         /// <param name="httpRequestMessage">message sended to server</param>
-        void OnSendingRequest(Uri uri, HttpMethod httpMethod, HttpRequestMessage httpRequestMessage);
+        /// <returns>A <see cref="Task"/></returns>
+        Task OnSendingRequestAsync(Uri uri, HttpMethod httpMethod, HttpRequestMessage httpRequestMessage);
 
         /// <summary>
         /// Invoked when received a response from server
@@ -28,7 +30,8 @@ namespace Tiny.RestClient
         /// <param name="httpMethod">verb of the request</param>
         /// <param name="response">response of the server</param>
         /// <param name="elapsedTime">time ellapsed between the send of request and response of server</param>
-        void OnReceivedResponse(Uri uri, HttpMethod httpMethod, HttpResponseMessage response, TimeSpan? elapsedTime);
+        /// <returns>A <see cref="Task"/></returns>
+        Task OnReceivedResponseAsync(Uri uri, HttpMethod httpMethod, HttpResponseMessage response, TimeSpan? elapsedTime);
 
         /// <summary>
         /// Invoke when a request failed to be invoked
@@ -37,6 +40,7 @@ namespace Tiny.RestClient
         /// <param name="httpMethod">verb of the request</param>
         /// <param name="exception">exception</param>
         /// <param name="elapsedTime">time ellapsed between the send of request and response of server (can be null if no listener measure time)</param>
-        void OnFailedToReceiveResponse(Uri uri, HttpMethod httpMethod, Exception exception, TimeSpan? elapsedTime);
+        /// <returns>A <see cref="Task"/></returns>
+        Task OnFailedToReceiveResponseAsync(Uri uri, HttpMethod httpMethod, Exception exception, TimeSpan? elapsedTime);
     }
 }