+++ Tiny.RestClient/Request/Content/IToSerializeContent.cs
@@ -1,10 +1,12 @@
-﻿using System.Text;
+﻿using System;
+using System.Text;
 
 namespace Tiny.RestClient
 {
     internal interface IToSerializeContent
     {
+        Type TypeToSerialize { get; }
         IFormatter Serializer { get; }
-        string GetSerializedStream(IFormatter serializer, Encoding encoding);
+        string GetSerializedString(IFormatter serializer, Encoding encoding);
     }
 }