+++ Tiny.RestClient/Request/MultipartFormContent/ToSerializeMultipartData.cs
@@ -1,4 +1,5 @@
-﻿using System.Text;
+﻿using System;
+using System.Text;
 
 namespace Tiny.RestClient
 {
@@ -10,9 +11,10 @@ namespace Tiny.RestClient
             Data = data;
         }
 
+        public Type TypeToSerialize => typeof(T);
         public T Data { get; }
 
-        public string GetSerializedStream(IFormatter serializer, Encoding encoding)
+        public string GetSerializedString(IFormatter serializer, Encoding encoding)
         {
             return serializer.Serialize<T>(Data, encoding);
         }