+++ Tiny.RestClient/TinyRestClient.cs
@@ -209,7 +209,7 @@ namespace Tiny.RestClient
 
                         if (formatter == null)
                         {
-                            if (response.Content.Headers?.ContentType.MediaType != null)
+                            if (response.Content.Headers?.ContentType?.MediaType != null)
                             {
                                 // TODO : optimize the seach of formatter ?
                                 // Try to find best formatter
@@ -441,7 +441,16 @@ namespace Tiny.RestClient
                 serializer = content.Serializer;
             }
 
-            var serializedString = content.GetSerializedStream(serializer, Settings.Encoding);
+            string serializedString = null;
+            try
+            {
+                serializedString = content.GetSerializedString(serializer, Settings.Encoding);
+            }
+            catch (Exception ex)
+            {
+                throw new SerializeException(content.TypeToSerialize, ex);
+            }
+
             if (serializedString == null)
             {
                 return null;
@@ -474,7 +483,7 @@ namespace Tiny.RestClient
 
         private void SetContentType(string contentType, HttpContent content)
         {
-            if (contentType == null)
+            if (string.IsNullOrEmpty(contentType))
             {
                 return;
             }
@@ -510,60 +519,63 @@ namespace Tiny.RestClient
                 stopwatch = new Stopwatch();
             }
 
-            try
+            using (var request = new HttpRequestMessage(httpMethod, uri))
             {
-                using (var request = new HttpRequestMessage(httpMethod, uri))
+                if (deserializer == null)
                 {
-                    if (deserializer == null)
-                    {
-                        deserializer = Settings.Formatters.Default;
-                    }
+                    deserializer = Settings.Formatters.Default;
+                }
 
+                if (!string.IsNullOrEmpty(deserializer.DefaultMediaType))
+                {
                     request.Headers.Accept.Add(new MediaTypeWithQualityHeaderValue(deserializer.DefaultMediaType));
+                }
 
-                    if (Settings.AddAcceptLanguageBasedOnCurrentCulture)
-                    {
-                        request.Headers.AcceptLanguage.Add(new StringWithQualityHeaderValue(CultureInfo.CurrentCulture.TwoLetterISOLanguageName));
-                    }
+                if (Settings.AddAcceptLanguageBasedOnCurrentCulture)
+                {
+                    request.Headers.AcceptLanguage.Add(new StringWithQualityHeaderValue(CultureInfo.CurrentCulture.TwoLetterISOLanguageName));
+                }
 
-                    foreach (var item in Settings.DefaultHeaders)
+                foreach (var item in Settings.DefaultHeaders)
+                {
+                    request.Headers.Add(item.Key, item.Value);
+                }
+
+                if (requestHeader != null)
+                {
+                    foreach (var item in requestHeader)
                     {
                         request.Headers.Add(item.Key, item.Value);
                     }
+                }
 
-                    if (requestHeader != null)
-                    {
-                        foreach (var item in requestHeader)
-                        {
-                            request.Headers.Add(item.Key, item.Value);
-                        }
-                    }
+                if (content != null)
+                {
+                    request.Content = content;
+                }
 
-                    if (content != null)
-                    {
-                        request.Content = content;
-                    }
-
-                    Settings.Listeners.OnSendingRequest(uri, httpMethod, request);
+                try
+                {
+                    await Settings.Listeners.OnSendingRequestAsync(uri, httpMethod, request).ConfigureAwait(false);
                     stopwatch?.Start();
                     var response = await _httpClient.SendAsync(request, HttpCompletionOption.ResponseContentRead, cancellationToken).ConfigureAwait(false);
 
                     stopwatch?.Stop();
-                    Settings.Listeners.OnReceivedResponse(uri, httpMethod, response, stopwatch?.Elapsed);
+                    await Settings.Listeners.OnReceivedResponseAsync(uri, httpMethod, response, stopwatch?.Elapsed).ConfigureAwait(false);
                     return response;
                 }
-            }
-            catch (Exception ex)
-            {
-                stopwatch?.Stop();
+                catch (Exception ex)
+                {
+                    stopwatch?.Stop();
 
-                Settings.Listeners.OnFailedToReceiveResponse(uri, httpMethod, ex, stopwatch?.Elapsed);
+                    await Settings.Listeners.OnFailedToReceiveResponseAsync(uri, httpMethod, ex, stopwatch?.Elapsed).ConfigureAwait(false);
 
-                throw new ConnectionException(
-                   "Failed to get a response from server",
-                   uri.AbsoluteUri,
-                   httpMethod.Method,
-                   ex);
+                    throw new ConnectionException(
+                       "Failed to get a response from server",
+                       uri.AbsoluteUri,
+                       httpMethod.Method,
+                       ex);
+                }
             }
         }